+++ src/extension.js
@@ -2290,6 +2290,123 @@ function activate(context) {
                 detail: "Allows columns to wrap when they exceed 100% of their parent’s width",
                 kind: vscode.CompletionItemKind.Class,
             },
+            {
+                label: "slds-table",
+                insertText: new vscode.SnippetString(
+                    "slds-table",
+                ),
+                detail: "Initializes data table",
+                documentation: "Can be used on table",
+                kind: vscode.CompletionItemKind.Class,
+            },
+            {
+                label: "slds-no-row-hover",
+                insertText: new vscode.SnippetString(
+                    "slds-no-row-hover",
+                ),
+                detail: "Default Table Row Hover",
+                documentation: "Can be used with .slds-table",
+                kind: vscode.CompletionItemKind.Class,
+            },
+            {
+                label: "slds-is-selected",
+                insertText: new vscode.SnippetString(
+                    "slds-is-selected",
+                ),
+                detail: "Selected Table Row + Hover",
+                documentation: "Can be used on .slds-table tr",
+                kind: vscode.CompletionItemKind.Class,
+            },
+            {
+                label: "slds-cell-wrap",
+                insertText: new vscode.SnippetString(
+                    "slds-cell-wrap",
+                ),
+                detail: "By default, nowrap is applied",
+                documentation: "Can be used on .slds-table th, .slds-table td",
+                kind: vscode.CompletionItemKind.Class,
+            },
+            {
+                label: "slds-cell-buffer_left",
+                insertText: new vscode.SnippetString(
+                    "slds-cell-buffer_left",
+                ),
+                detail: "Use to add a left padding buffer to cells",
+                documentation: "Can be used on .slds-table th, .slds-table td",
+                kind: vscode.CompletionItemKind.Class,
+            },
+            {
+                label: "slds-cell-buffer_right",
+                insertText: new vscode.SnippetString(
+                    "slds-cell-buffer_right",
+                ),
+                detail: "Use to add a right padding buffer to cells",
+                documentation: "Can be used on .slds-table th, .slds-table td",
+                kind: vscode.CompletionItemKind.Class,
+            },
+            {
+                label: "slds-table_cell-buffer",
+                insertText: new vscode.SnippetString(
+                    "slds-table_cell-buffer",
+                ),
+                detail: "Add left and right padding to the first and last cells of a table",
+                documentation: "Can be used with .slds-table",
+                kind: vscode.CompletionItemKind.Class,
+            },
+            {
+                label: "slds-table_bordered",
+                insertText: new vscode.SnippetString(
+                    "slds-table_bordered",
+                ),
+                detail: "Add vertical borders to a table",
+                documentation: "Can be used with .slds-table",
+                kind: vscode.CompletionItemKind.Class,
+            },
+            {
+                label: "slds-has-focus",
+                insertText: new vscode.SnippetString(
+                    "slds-has-focus",
+                ),
+                detail: "Focus state on a cell",
+                documentation: "Can be on .slds-table th, .slds-table td",
+                kind: vscode.CompletionItemKind.Class,
+            },
+            {
+                label: "slds-table_bordered",
+                insertText: new vscode.SnippetString(
+                    "slds-table_bordered",
+                ),
+                detail: "Add vertical borders to a table",
+                documentation: "Can be used with .slds-table",
+                kind: vscode.CompletionItemKind.Class,
+            },
+            {
+                label: "slds-table_col-bordered",
+                insertText: new vscode.SnippetString(
+                    "slds-table_col-bordered",
+                ),
+                detail: "Add column borders",
+                documentation: "Can be used with .slds-table",
+                kind: vscode.CompletionItemKind.Class,
+            },
+            {
+                label: "slds-table_striped",
+                insertText: new vscode.SnippetString(
+                    "slds-table_striped",
+                ),
+                detail: "Add alternating strips to rows",
+                documentation: "Can be used with .slds-table",
+                kind: vscode.CompletionItemKind.Class,
+            },
+            {
+                label: "slds-table_fixed-layout",
+                insertText: new vscode.SnippetString(
+                    "slds-table_fixed-layout",
+                ),
+                detail: "Set table to use fixed layout for width and truncation purposes",
+                documentation: "Can be used with .slds-table",
+                kind: vscode.CompletionItemKind.Class,
+            },
         ];
 
         if( isAdvancedMode ) {
@@ -3809,6 +3926,14 @@ function activate(context) {
                 kind: vscode.CompletionItemKind.Snippet,
             },
             {
+                label: "input:checkbox-aura",
+                insertText: new vscode.SnippetString(
+                    "<lightning:input type=\"checkbox\" label=\"${1:label}\" name=\"${2:name}\" checked=\"${3|true,false|}\" required=\"${4|true,false|}\" disabled=\"${5|true,false|}\"/>"
+                ),
+                detail: "Checkbox options can be required or disabled.",
+                kind: vscode.CompletionItemKind.Snippet,
+            },
+            {
                 label: "combobox-aura",
                 insertText: new vscode.SnippetString(
                     "<aura:attribute name=\"${1:Options}\" type=\"List\" description=\"${2:description}\"/>\n<lightning:combobox name=\"${3:name}\" label=\"${4:label}\" placeholder=\"${5:placeholder}\" options=\"{! v.${1:Options} }\" onchange=\"{! c.${6:function} }\"/>"
@@ -3827,7 +3952,7 @@ function activate(context) {
             {
                 label: "icon:aura",
                 insertText: new vscode.SnippetString(
-                    "<lightning:icon iconName=\"${1|utility:activity, utility:ad_set, utility:add, utility:adduser, utility:advanced_function, utility:advertising, utility:agent_session, utility:alert, utility:all, utility:anchor, utility:animal_and_nature, utility:announcement, utility:answer, utility:answered_twice, utility:apex_plugin, utility:apex, utility:approval, utility:apps, utility:archive, utility:arrowdown, utility:arrowup, utility:assignment, utility:attach, utility:automate, utility:away, utility:back, utility:ban, utility:block_visitor, utility:bold, utility:bookmark, utility:breadcrumbs, utility:broadcast, utility:brush, utility:bucket, utility:builder, utility:button_choice, utility:call, utility:campaign, utility:cancel_file_request, utility:cancel_transfer, utility:capslock, utility:cart, utility:case, utility:cases, utility:center_align_text, utility:change_owner, utility:change_record_type, utility:chart, utility:chat, utility:check, utility:checkin, utility:chevrondown, utility:chevronleft, utility:chevronright, utility:chevronup, utility:choice, utility:classic_interface, utility:clear, utility:clock, utility:close, utility:collapse_all, utility:collection_variable, utility:color_swatch, utility:comments, utility:company, utility:component_customization, utility:connected_apps, utility:constant, utility:contact_request, utility:contract_alt, utility:contract, utility:copy_to_clipboard, utility:copy, utility:crossfilter, utility:currency_input, utility:currency, utility:custom_apps, utility:cut, utility:dash, utility:database, utility:datadotcom, utility:date_input, utility:date_time, utility:dayview, utility:delete, utility:deprecate, utility:description, utility:desktop_and_phone, utility:desktop_console, utility:desktop, utility:dialing, utility:diamond, utility:dislike, utility:display_rich_text, utility:display_text, utility:dock_panel, utility:down, utility:download, utility:drag_and_drop, utility:drag, utility:dynamic_record_choice, utility:edit_form, utility:edit, utility:education, utility:einstein, utility:email_open, utility:email, utility:emoji, utility:end_call, utility:end_chat, utility:end_messaging_session, utility:engage, utility:enter, utility:erect_window, utility:error, utility:event, utility:events, utility:expand_all, utility:expand_alt, utility:expand, utility:fallback, utility:favorite, utility:feed, utility:file, utility:filter, utility:filterList, utility:flow_alt, utility:flow, utility:food_and_drink, utility:formula, utility:forward_up, utility:forward, utility:frozen, utility:fulfillment_order, utility:full_width_view, utility:global_constant, utility:graph, utility:groups, utility:help_center, utility:help, utility:hide_mobile, utility:hide, utility:hierarchy, utility:high_velocity_sales, utility:home, utility:identity, utility:image, utility:in_app_assistant, utility:inbox, utility:incoming_call, utility:info_alt, utility:info, utility:insert_tag_field, utility:insert_template, utility:inspector_panel, utility:internal_share, utility:italic, utility:jump_to_bottom, utility:jump_to_top, utility:justify_text, utility:kanban, utility:keyboard_dismiss, utility:knowledge_base, utility:layers, utility:layout, utility:leave_conference, utility:left_align_text, utility:left, utility:level_down, utility:level_up, utility:light_bulb, utility:lightning_extension, utility:lightning_inspector, utility:like, utility:link, utility:linked, utility:list, utility:listen, utility:live_message, utility:location, utility:lock, utility:locker_service_api_viewer, utility:locker_service_console, utility:log_a_call, utility:logout, utility:loop, utility:lower_flag, utility:macros, utility:magicwand, utility:mark_all_as_read, utility:matrix, utility:merge_field, utility:merge, utility:metrics, utility:minimize_window, utility:missed_call, utility:money, utility:moneybag, utility:monthlyview, utility:move, utility:multi_picklist, utility:multi_select_checkbox, utility:muted, utility:new_direct_message, utility:new_window, utility:new, utility:news, utility:note, utility:notebook, utility:notification, utility:number_input, utility:office365, utility:offline_cached, utility:offline, utility:omni_channel, utility:open_folder, utility:open, utility:opened_folder, utility:outbound_call, utility:outcome, utility:overflow, utility:package_org_beta, utility:package_org, utility:package, utility:page, utility:palette, utility:password, utility:paste, utility:pause, utility:people, utility:phone_landscape, utility:phone_portrait, utility:photo, utility:picklist_choice, utility:picklist_type, utility:picklist, utility:pin, utility:pinned, utility:play, utility:podcast_webinar, utility:pop_in, utility:power, utility:preview, utility:print, utility:priority, utility:privately_shared, utility:process, utility:prompt_edit, utility:prompt, utility:push, utility:puzzle, utility:question_mark, utility:question, utility:questions_and_answers, utility:quick_text, utility:quip, utility:quotation_marks, utility:quote, utility:radio_button, utility:rating, utility:reassign, utility:record_create, utility:record_delete, utility:record_lookup, utility:record_update, utility:record, utility:recurring_exception, utility:recycle_bin_empty, utility:recycle_bin_full, utility:redo, utility:refresh, utility:relate, utility:reminder, utility:remove_formatting, utility:remove_link, utility:replace, utility:reply_all, utility:reply, utility:report_issue, utility:reset_password, utility:resource_absence, utility:resource_capacity, utility:resource_territory, utility:retail_execution, utility:retweet, utility:ribbon, utility:richtextbulletedlist, utility:richtextindent, utility:richtextnumberedlist, utility:richtextoutdent, utility:right_align_text, utility:right, utility:rotate, utility:routing_offline, utility:rows, utility:rules, utility:salesforce1, utility:save, utility:screen, utility:search, utility:section, utility:send, utility:sentiment_negative, utility:sentiment_neutral, utility:settings, utility:setup_assistant_guide, utility:setup_modal, utility:setup, utility:share_file, utility:share_mobile, utility:share_post, utility:share, utility:shield, utility:shift_ui, utility:shopping_bag, utility:shortcuts, utility:side_list, utility:signpost, utility:skip_back, utility:skip_forward, utility:skip, utility:slider, utility:smiley_and_people, utility:sms, utility:snippet, utility:sobject_collection, utility:sobject, utility:socialshare, utility:sort, utility:spinner, utility:stage_collection, utility:stage, utility:standard_objects, utility:steps, utility:stop, utility:store, utility:strategy, utility:strikethrough, utility:success, utility:summary, utility:summarydetail, utility:survey, utility:switch, utility:symbols, utility:sync, utility:system_and_global_variable, utility:table_settings, utility:table, utility:tablet_landscape, utility:tablet_portrait, utility:tabset, utility:target, utility:task, utility:text_background_color, utility:text_color, utility:text_template, utility:text, utility:textarea, utility:textbox, utility:threedots_vertical, utility:threedots, utility:thunder, utility:tile_card_list, utility:toggle_panel_bottom, utility:toggle_panel_left, utility:toggle_panel_right, utility:toggle_panel_top, utility:toggle, utility:topic, utility:topic2, utility:touch_action, utility:tracker, utility:trail, utility:trailhead, utility:travel_and_places, utility:trending, utility:turn_off_notifications, utility:type_tool, utility:type, utility:undelete, utility:undeprecate, utility:underline, utility:undo, utility:unlinked, utility:unlock, utility:unmuted, utility:up, utility:upload, utility:user_role, utility:user, utility:variable, utility:video, utility:voicemail_drop, utility:volume_high, utility:volume_low, utility:volume_off, utility:waits, utility:warning, utility:watchlist, utility:weeklyview, utility:wifi, utility:work_order_type, utility:world, utility:yubi_key, utility:zoomin, utility:zoomout|}\" alternativeText=\"${2:Title}\" title=\"${2:Title}\" />",
+                    "<lightning:icon iconName=\"${1|utility:activity,utility:ad_set,utility:add,utility:adduser,utility:advanced_function,utility:advertising,utility:agent_session,utility:alert,utility:all,utility:anchor,utility:animal_and_nature,utility:announcement,utility:answer,utility:answered_twice,utility:apex_plugin,utility:apex,utility:approval,utility:apps,utility:archive,utility:arrowdown,utility:arrowup,utility:assignment,utility:attach,utility:automate,utility:away,utility:back,utility:ban,utility:block_visitor,utility:bold,utility:bookmark,utility:breadcrumbs,utility:broadcast,utility:brush,utility:bucket,utility:builder,utility:button_choice,utility:call,utility:campaign,utility:cancel_file_request,utility:cancel_transfer,utility:capslock,utility:cart,utility:case,utility:cases,utility:center_align_text,utility:change_owner,utility:change_record_type,utility:chart,utility:chat,utility:check,utility:checkin,utility:chevrondown,utility:chevronleft,utility:chevronright,utility:chevronup,utility:choice,utility:classic_interface,utility:clear,utility:clock,utility:close,utility:collapse_all,utility:collection_variable,utility:color_swatch,utility:comments,utility:company,utility:component_customization,utility:connected_apps,utility:constant,utility:contact_request,utility:contract_alt,utility:contract,utility:copy_to_clipboard,utility:copy,utility:crossfilter,utility:currency_input,utility:currency,utility:custom_apps,utility:cut,utility:dash,utility:database,utility:datadotcom,utility:date_input,utility:date_time,utility:dayview,utility:delete,utility:deprecate,utility:description,utility:desktop_and_phone,utility:desktop_console,utility:desktop,utility:dialing,utility:diamond,utility:dislike,utility:display_rich_text,utility:display_text,utility:dock_panel,utility:down,utility:download,utility:drag_and_drop,utility:drag,utility:dynamic_record_choice,utility:edit_form,utility:edit,utility:education,utility:einstein,utility:email_open,utility:email,utility:emoji,utility:end_call,utility:end_chat,utility:end_messaging_session,utility:engage,utility:enter,utility:erect_window,utility:error,utility:event,utility:events,utility:expand_all,utility:expand_alt,utility:expand,utility:fallback,utility:favorite,utility:feed,utility:file,utility:filter,utility:filterList,utility:flow_alt,utility:flow,utility:food_and_drink,utility:formula,utility:forward_up,utility:forward,utility:frozen,utility:fulfillment_order,utility:full_width_view,utility:global_constant,utility:graph,utility:groups,utility:help_center,utility:help,utility:hide_mobile,utility:hide,utility:hierarchy,utility:high_velocity_sales,utility:home,utility:identity,utility:image,utility:in_app_assistant,utility:inbox,utility:incoming_call,utility:info_alt,utility:info,utility:insert_tag_field,utility:insert_template,utility:inspector_panel,utility:internal_share,utility:italic,utility:jump_to_bottom,utility:jump_to_top,utility:justify_text,utility:kanban,utility:keyboard_dismiss,utility:knowledge_base,utility:layers,utility:layout,utility:leave_conference,utility:left_align_text,utility:left,utility:level_down,utility:level_up,utility:light_bulb,utility:lightning_extension,utility:lightning_inspector,utility:like,utility:link,utility:linked,utility:list,utility:listen,utility:live_message,utility:location,utility:lock,utility:locker_service_api_viewer,utility:locker_service_console,utility:log_a_call,utility:logout,utility:loop,utility:lower_flag,utility:macros,utility:magicwand,utility:mark_all_as_read,utility:matrix,utility:merge_field,utility:merge,utility:metrics,utility:minimize_window,utility:missed_call,utility:money,utility:moneybag,utility:monthlyview,utility:move,utility:multi_picklist,utility:multi_select_checkbox,utility:muted,utility:new_direct_message,utility:new_window,utility:new,utility:news,utility:note,utility:notebook,utility:notification,utility:number_input,utility:office365,utility:offline_cached,utility:offline,utility:omni_channel,utility:open_folder,utility:open,utility:opened_folder,utility:outbound_call,utility:outcome,utility:overflow,utility:package_org_beta,utility:package_org,utility:package,utility:page,utility:palette,utility:password,utility:paste,utility:pause,utility:people,utility:phone_landscape,utility:phone_portrait,utility:photo,utility:picklist_choice,utility:picklist_type,utility:picklist,utility:pin,utility:pinned,utility:play,utility:podcast_webinar,utility:pop_in,utility:power,utility:preview,utility:print,utility:priority,utility:privately_shared,utility:process,utility:prompt_edit,utility:prompt,utility:push,utility:puzzle,utility:question_mark,utility:question,utility:questions_and_answers,utility:quick_text,utility:quip,utility:quotation_marks,utility:quote,utility:radio_button,utility:rating,utility:reassign,utility:record_create,utility:record_delete,utility:record_lookup,utility:record_update,utility:record,utility:recurring_exception,utility:recycle_bin_empty,utility:recycle_bin_full,utility:redo,utility:refresh,utility:relate,utility:reminder,utility:remove_formatting,utility:remove_link,utility:replace,utility:reply_all,utility:reply,utility:report_issue,utility:reset_password,utility:resource_absence,utility:resource_capacity,utility:resource_territory,utility:retail_execution,utility:retweet,utility:ribbon,utility:richtextbulletedlist,utility:richtextindent,utility:richtextnumberedlist,utility:richtextoutdent,utility:right_align_text,utility:right,utility:rotate,utility:routing_offline,utility:rows,utility:rules,utility:salesforce1,utility:save,utility:screen,utility:search,utility:section,utility:send,utility:sentiment_negative,utility:sentiment_neutral,utility:settings,utility:setup_assistant_guide,utility:setup_modal,utility:setup,utility:share_file,utility:share_mobile,utility:share_post,utility:share,utility:shield,utility:shift_ui,utility:shopping_bag,utility:shortcuts,utility:side_list,utility:signpost,utility:skip_back,utility:skip_forward,utility:skip,utility:slider,utility:smiley_and_people,utility:sms,utility:snippet,utility:sobject_collection,utility:sobject,utility:socialshare,utility:sort,utility:spinner,utility:stage_collection,utility:stage,utility:standard_objects,utility:steps,utility:stop,utility:store,utility:strategy,utility:strikethrough,utility:success,utility:summary,utility:summarydetail,utility:survey,utility:switch,utility:symbols,utility:sync,utility:system_and_global_variable,utility:table_settings,utility:table,utility:tablet_landscape,utility:tablet_portrait,utility:tabset,utility:target,utility:task,utility:text_background_color,utility:text_color,utility:text_template,utility:text,utility:textarea,utility:textbox,utility:threedots_vertical,utility:threedots,utility:thunder,utility:tile_card_list,utility:toggle_panel_bottom,utility:toggle_panel_left,utility:toggle_panel_right,utility:toggle_panel_top,utility:toggle,utility:topic,utility:topic2,utility:touch_action,utility:tracker,utility:trail,utility:trailhead,utility:travel_and_places,utility:trending,utility:turn_off_notifications,utility:type_tool,utility:type,utility:undelete,utility:undeprecate,utility:underline,utility:undo,utility:unlinked,utility:unlock,utility:unmuted,utility:up,utility:upload,utility:user_role,utility:user,utility:variable,utility:video,utility:voicemail_drop,utility:volume_high,utility:volume_low,utility:volume_off,utility:waits,utility:warning,utility:watchlist,utility:weeklyview,utility:wifi,utility:work_order_type,utility:world,utility:yubi_key,utility:zoomin,utility:zoomout|}\" alternativeText=\"${2:Title}\" title=\"${2:Title}\" variant=\"${3|Success,warning,error,inverse|}\" size=\"${4|xx-small,x-small,small,medium,large|}\" />",
                 ),
                 detail: "Aura Icon",
                 kind: vscode.CompletionItemKind.Snippet,
@@ -3859,9 +3984,57 @@ function activate(context) {
             {
                 label: "if",
                 insertText: new vscode.SnippetString(
-                    "<aura:if isTrue=\"{!v.${1:bar}}\">\n\t${2:ifBody}\n\t<aura:set attribute=\"else\">\n\t\t${3:elseBody}\n\t</aura:set>\n</aura:if>",
+                    "<aura:if isTrue=\"{!v.${1:variable}}\">\n\t${2:ifBody}\n</aura:if>",
+                    ),
+                detail: "Conditionally instantiates and renders either the body.",
+                kind: vscode.CompletionItemKind.Method,
+            },
+            {
+                label: "if empty",
+                insertText: new vscode.SnippetString(
+                    "<aura:if isTrue=\"{!empty(v.${1:variable})}\">\n\t${2:ifBody}\n</aura:if>",
+                    ),
+                detail: "Conditionally checks if empty and renders either the body.",
+                kind: vscode.CompletionItemKind.Method,
+            },
+            {
+                label: "if not empty",
+                insertText: new vscode.SnippetString(
+                    "<aura:if isTrue=\"{!not(empty(v.${1:variable}))}\">\n\t${2:ifBody}\n</aura:if>",
+                    ),
+                detail: "Conditionally checks if not empty and renders either the body.",
+                kind: vscode.CompletionItemKind.Method,
+            },
+            {
+                label: "if equals",
+                insertText: new vscode.SnippetString(
+                    "<aura:if isTrue=\"{!v.${1:variable} == '${2:text}'}\">\n\t${3:ifBody}\n</aura:if>",
+                    ),
+                detail: "Conditionally checks if equals and renders either the body.",
+                kind: vscode.CompletionItemKind.Method,
+            },
+            {
+                label: "if not equals",
+                insertText: new vscode.SnippetString(
+                    "<aura:if isTrue=\"{!not(v.${1:variable} == '${2:text}')}\">\n\t${3:ifBody}\n</aura:if>",
+                    ),
+                detail: "Conditionally if not equals and renders either the body.",
+                kind: vscode.CompletionItemKind.Method,
+            },
+            {
+                label: "if else",
+                insertText: new vscode.SnippetString(
+                    "<aura:if isTrue=\"{!v.${1:variable}}\">\n\t${2:ifBody}\n\t<aura:set attribute=\"else\">\n\t\t${3:elseBody}\n\t</aura:set>\n</aura:if>",
                     ),
                 detail: "Conditionally instantiates and renders either the body or the components in the else attribute.",
+                kind: vscode.CompletionItemKind.Method,
+            },
+            {
+                label: "else",
+                insertText: new vscode.SnippetString(
+                    "<aura:set attribute=\"else\">\n\t${3:elseBody}\n</aura:set>",
+                    ),
+                detail: "Renders else attribute",
                 kind: vscode.CompletionItemKind.Method,
             },
             {
@@ -4022,6 +4195,14 @@ function activate(context) {
                 kind: vscode.CompletionItemKind.Snippet,
             },
             {
+                label: "input:checkbox-lwc",
+                insertText: new vscode.SnippetString(
+                    "<lightning-input type=\"checkbox\" label=\"${1:label}\" name=\"${2:name}\" ${3:checked} ${4:required} ${5:disabled}></lightning-input>"
+                ),
+                detail: "Checkbox options can be required or disabled.",
+                kind: vscode.CompletionItemKind.Snippet,
+            },
+            {
                 label: "combobox-lwc",
                 insertText: new vscode.SnippetString(
                     "<lightning-combobox\n\tname=\"${1:name}\"\n\tlabel=\"${2:label}\"\n\tvalue=\"\"\n\tplaceholder=\"${3:placeholder}\"\n\toptions={options}\n\tonchange={handleChange}\n\trequired\n></lightning-combobox>"
@@ -4040,7 +4221,7 @@ function activate(context) {
             {
                 label: "icon:lwc",
                 insertText: new vscode.SnippetString(
-                    "<lightning-icon icon-name=\"${1|utility:activity, utility:ad_set, utility:add, utility:adduser, utility:advanced_function, utility:advertising, utility:agent_session, utility:alert, utility:all, utility:anchor, utility:animal_and_nature, utility:announcement, utility:answer, utility:answered_twice, utility:apex_plugin, utility:apex, utility:approval, utility:apps, utility:archive, utility:arrowdown, utility:arrowup, utility:assignment, utility:attach, utility:automate, utility:away, utility:back, utility:ban, utility:block_visitor, utility:bold, utility:bookmark, utility:breadcrumbs, utility:broadcast, utility:brush, utility:bucket, utility:builder, utility:button_choice, utility:call, utility:campaign, utility:cancel_file_request, utility:cancel_transfer, utility:capslock, utility:cart, utility:case, utility:cases, utility:center_align_text, utility:change_owner, utility:change_record_type, utility:chart, utility:chat, utility:check, utility:checkin, utility:chevrondown, utility:chevronleft, utility:chevronright, utility:chevronup, utility:choice, utility:classic_interface, utility:clear, utility:clock, utility:close, utility:collapse_all, utility:collection_variable, utility:color_swatch, utility:comments, utility:company, utility:component_customization, utility:connected_apps, utility:constant, utility:contact_request, utility:contract_alt, utility:contract, utility:copy_to_clipboard, utility:copy, utility:crossfilter, utility:currency_input, utility:currency, utility:custom_apps, utility:cut, utility:dash, utility:database, utility:datadotcom, utility:date_input, utility:date_time, utility:dayview, utility:delete, utility:deprecate, utility:description, utility:desktop_and_phone, utility:desktop_console, utility:desktop, utility:dialing, utility:diamond, utility:dislike, utility:display_rich_text, utility:display_text, utility:dock_panel, utility:down, utility:download, utility:drag_and_drop, utility:drag, utility:dynamic_record_choice, utility:edit_form, utility:edit, utility:education, utility:einstein, utility:email_open, utility:email, utility:emoji, utility:end_call, utility:end_chat, utility:end_messaging_session, utility:engage, utility:enter, utility:erect_window, utility:error, utility:event, utility:events, utility:expand_all, utility:expand_alt, utility:expand, utility:fallback, utility:favorite, utility:feed, utility:file, utility:filter, utility:filterList, utility:flow_alt, utility:flow, utility:food_and_drink, utility:formula, utility:forward_up, utility:forward, utility:frozen, utility:fulfillment_order, utility:full_width_view, utility:global_constant, utility:graph, utility:groups, utility:help_center, utility:help, utility:hide_mobile, utility:hide, utility:hierarchy, utility:high_velocity_sales, utility:home, utility:identity, utility:image, utility:in_app_assistant, utility:inbox, utility:incoming_call, utility:info_alt, utility:info, utility:insert_tag_field, utility:insert_template, utility:inspector_panel, utility:internal_share, utility:italic, utility:jump_to_bottom, utility:jump_to_top, utility:justify_text, utility:kanban, utility:keyboard_dismiss, utility:knowledge_base, utility:layers, utility:layout, utility:leave_conference, utility:left_align_text, utility:left, utility:level_down, utility:level_up, utility:light_bulb, utility:lightning_extension, utility:lightning_inspector, utility:like, utility:link, utility:linked, utility:list, utility:listen, utility:live_message, utility:location, utility:lock, utility:locker_service_api_viewer, utility:locker_service_console, utility:log_a_call, utility:logout, utility:loop, utility:lower_flag, utility:macros, utility:magicwand, utility:mark_all_as_read, utility:matrix, utility:merge_field, utility:merge, utility:metrics, utility:minimize_window, utility:missed_call, utility:money, utility:moneybag, utility:monthlyview, utility:move, utility:multi_picklist, utility:multi_select_checkbox, utility:muted, utility:new_direct_message, utility:new_window, utility:new, utility:news, utility:note, utility:notebook, utility:notification, utility:number_input, utility:office365, utility:offline_cached, utility:offline, utility:omni_channel, utility:open_folder, utility:open, utility:opened_folder, utility:outbound_call, utility:outcome, utility:overflow, utility:package_org_beta, utility:package_org, utility:package, utility:page, utility:palette, utility:password, utility:paste, utility:pause, utility:people, utility:phone_landscape, utility:phone_portrait, utility:photo, utility:picklist_choice, utility:picklist_type, utility:picklist, utility:pin, utility:pinned, utility:play, utility:podcast_webinar, utility:pop_in, utility:power, utility:preview, utility:print, utility:priority, utility:privately_shared, utility:process, utility:prompt_edit, utility:prompt, utility:push, utility:puzzle, utility:question_mark, utility:question, utility:questions_and_answers, utility:quick_text, utility:quip, utility:quotation_marks, utility:quote, utility:radio_button, utility:rating, utility:reassign, utility:record_create, utility:record_delete, utility:record_lookup, utility:record_update, utility:record, utility:recurring_exception, utility:recycle_bin_empty, utility:recycle_bin_full, utility:redo, utility:refresh, utility:relate, utility:reminder, utility:remove_formatting, utility:remove_link, utility:replace, utility:reply_all, utility:reply, utility:report_issue, utility:reset_password, utility:resource_absence, utility:resource_capacity, utility:resource_territory, utility:retail_execution, utility:retweet, utility:ribbon, utility:richtextbulletedlist, utility:richtextindent, utility:richtextnumberedlist, utility:richtextoutdent, utility:right_align_text, utility:right, utility:rotate, utility:routing_offline, utility:rows, utility:rules, utility:salesforce1, utility:save, utility:screen, utility:search, utility:section, utility:send, utility:sentiment_negative, utility:sentiment_neutral, utility:settings, utility:setup_assistant_guide, utility:setup_modal, utility:setup, utility:share_file, utility:share_mobile, utility:share_post, utility:share, utility:shield, utility:shift_ui, utility:shopping_bag, utility:shortcuts, utility:side_list, utility:signpost, utility:skip_back, utility:skip_forward, utility:skip, utility:slider, utility:smiley_and_people, utility:sms, utility:snippet, utility:sobject_collection, utility:sobject, utility:socialshare, utility:sort, utility:spinner, utility:stage_collection, utility:stage, utility:standard_objects, utility:steps, utility:stop, utility:store, utility:strategy, utility:strikethrough, utility:success, utility:summary, utility:summarydetail, utility:survey, utility:switch, utility:symbols, utility:sync, utility:system_and_global_variable, utility:table_settings, utility:table, utility:tablet_landscape, utility:tablet_portrait, utility:tabset, utility:target, utility:task, utility:text_background_color, utility:text_color, utility:text_template, utility:text, utility:textarea, utility:textbox, utility:threedots_vertical, utility:threedots, utility:thunder, utility:tile_card_list, utility:toggle_panel_bottom, utility:toggle_panel_left, utility:toggle_panel_right, utility:toggle_panel_top, utility:toggle, utility:topic, utility:topic2, utility:touch_action, utility:tracker, utility:trail, utility:trailhead, utility:travel_and_places, utility:trending, utility:turn_off_notifications, utility:type_tool, utility:type, utility:undelete, utility:undeprecate, utility:underline, utility:undo, utility:unlinked, utility:unlock, utility:unmuted, utility:up, utility:upload, utility:user_role, utility:user, utility:variable, utility:video, utility:voicemail_drop, utility:volume_high, utility:volume_low, utility:volume_off, utility:waits, utility:warning, utility:watchlist, utility:weeklyview, utility:wifi, utility:work_order_type, utility:world, utility:yubi_key, utility:zoomin, utility:zoomout|}\" alternative-text=\"${2:Title}\" title=\"${2:Title}\"></lightning-icon>",
+                    "<lightning-icon icon-name=\"${1|utility:activity,utility:ad_set,utility:add,utility:adduser,utility:advanced_function,utility:advertising,utility:agent_session,utility:alert,utility:all,utility:anchor,utility:animal_and_nature,utility:announcement,utility:answer,utility:answered_twice,utility:apex_plugin,utility:apex,utility:approval,utility:apps,utility:archive,utility:arrowdown,utility:arrowup,utility:assignment,utility:attach,utility:automate,utility:away,utility:back,utility:ban,utility:block_visitor,utility:bold,utility:bookmark,utility:breadcrumbs,utility:broadcast,utility:brush,utility:bucket,utility:builder,utility:button_choice,utility:call,utility:campaign,utility:cancel_file_request,utility:cancel_transfer,utility:capslock,utility:cart,utility:case,utility:cases,utility:center_align_text,utility:change_owner,utility:change_record_type,utility:chart,utility:chat,utility:check,utility:checkin,utility:chevrondown,utility:chevronleft,utility:chevronright,utility:chevronup,utility:choice,utility:classic_interface,utility:clear,utility:clock,utility:close,utility:collapse_all,utility:collection_variable,utility:color_swatch,utility:comments,utility:company,utility:component_customization,utility:connected_apps,utility:constant,utility:contact_request,utility:contract_alt,utility:contract,utility:copy_to_clipboard,utility:copy,utility:crossfilter,utility:currency_input,utility:currency,utility:custom_apps,utility:cut,utility:dash,utility:database,utility:datadotcom,utility:date_input,utility:date_time,utility:dayview,utility:delete,utility:deprecate,utility:description,utility:desktop_and_phone,utility:desktop_console,utility:desktop,utility:dialing,utility:diamond,utility:dislike,utility:display_rich_text,utility:display_text,utility:dock_panel,utility:down,utility:download,utility:drag_and_drop,utility:drag,utility:dynamic_record_choice,utility:edit_form,utility:edit,utility:education,utility:einstein,utility:email_open,utility:email,utility:emoji,utility:end_call,utility:end_chat,utility:end_messaging_session,utility:engage,utility:enter,utility:erect_window,utility:error,utility:event,utility:events,utility:expand_all,utility:expand_alt,utility:expand,utility:fallback,utility:favorite,utility:feed,utility:file,utility:filter,utility:filterList,utility:flow_alt,utility:flow,utility:food_and_drink,utility:formula,utility:forward_up,utility:forward,utility:frozen,utility:fulfillment_order,utility:full_width_view,utility:global_constant,utility:graph,utility:groups,utility:help_center,utility:help,utility:hide_mobile,utility:hide,utility:hierarchy,utility:high_velocity_sales,utility:home,utility:identity,utility:image,utility:in_app_assistant,utility:inbox,utility:incoming_call,utility:info_alt,utility:info,utility:insert_tag_field,utility:insert_template,utility:inspector_panel,utility:internal_share,utility:italic,utility:jump_to_bottom,utility:jump_to_top,utility:justify_text,utility:kanban,utility:keyboard_dismiss,utility:knowledge_base,utility:layers,utility:layout,utility:leave_conference,utility:left_align_text,utility:left,utility:level_down,utility:level_up,utility:light_bulb,utility:lightning_extension,utility:lightning_inspector,utility:like,utility:link,utility:linked,utility:list,utility:listen,utility:live_message,utility:location,utility:lock,utility:locker_service_api_viewer,utility:locker_service_console,utility:log_a_call,utility:logout,utility:loop,utility:lower_flag,utility:macros,utility:magicwand,utility:mark_all_as_read,utility:matrix,utility:merge_field,utility:merge,utility:metrics,utility:minimize_window,utility:missed_call,utility:money,utility:moneybag,utility:monthlyview,utility:move,utility:multi_picklist,utility:multi_select_checkbox,utility:muted,utility:new_direct_message,utility:new_window,utility:new,utility:news,utility:note,utility:notebook,utility:notification,utility:number_input,utility:office365,utility:offline_cached,utility:offline,utility:omni_channel,utility:open_folder,utility:open,utility:opened_folder,utility:outbound_call,utility:outcome,utility:overflow,utility:package_org_beta,utility:package_org,utility:package,utility:page,utility:palette,utility:password,utility:paste,utility:pause,utility:people,utility:phone_landscape,utility:phone_portrait,utility:photo,utility:picklist_choice,utility:picklist_type,utility:picklist,utility:pin,utility:pinned,utility:play,utility:podcast_webinar,utility:pop_in,utility:power,utility:preview,utility:print,utility:priority,utility:privately_shared,utility:process,utility:prompt_edit,utility:prompt,utility:push,utility:puzzle,utility:question_mark,utility:question,utility:questions_and_answers,utility:quick_text,utility:quip,utility:quotation_marks,utility:quote,utility:radio_button,utility:rating,utility:reassign,utility:record_create,utility:record_delete,utility:record_lookup,utility:record_update,utility:record,utility:recurring_exception,utility:recycle_bin_empty,utility:recycle_bin_full,utility:redo,utility:refresh,utility:relate,utility:reminder,utility:remove_formatting,utility:remove_link,utility:replace,utility:reply_all,utility:reply,utility:report_issue,utility:reset_password,utility:resource_absence,utility:resource_capacity,utility:resource_territory,utility:retail_execution,utility:retweet,utility:ribbon,utility:richtextbulletedlist,utility:richtextindent,utility:richtextnumberedlist,utility:richtextoutdent,utility:right_align_text,utility:right,utility:rotate,utility:routing_offline,utility:rows,utility:rules,utility:salesforce1,utility:save,utility:screen,utility:search,utility:section,utility:send,utility:sentiment_negative,utility:sentiment_neutral,utility:settings,utility:setup_assistant_guide,utility:setup_modal,utility:setup,utility:share_file,utility:share_mobile,utility:share_post,utility:share,utility:shield,utility:shift_ui,utility:shopping_bag,utility:shortcuts,utility:side_list,utility:signpost,utility:skip_back,utility:skip_forward,utility:skip,utility:slider,utility:smiley_and_people,utility:sms,utility:snippet,utility:sobject_collection,utility:sobject,utility:socialshare,utility:sort,utility:spinner,utility:stage_collection,utility:stage,utility:standard_objects,utility:steps,utility:stop,utility:store,utility:strategy,utility:strikethrough,utility:success,utility:summary,utility:summarydetail,utility:survey,utility:switch,utility:symbols,utility:sync,utility:system_and_global_variable,utility:table_settings,utility:table,utility:tablet_landscape,utility:tablet_portrait,utility:tabset,utility:target,utility:task,utility:text_background_color,utility:text_color,utility:text_template,utility:text,utility:textarea,utility:textbox,utility:threedots_vertical,utility:threedots,utility:thunder,utility:tile_card_list,utility:toggle_panel_bottom,utility:toggle_panel_left,utility:toggle_panel_right,utility:toggle_panel_top,utility:toggle,utility:topic,utility:topic2,utility:touch_action,utility:tracker,utility:trail,utility:trailhead,utility:travel_and_places,utility:trending,utility:turn_off_notifications,utility:type_tool,utility:type,utility:undelete,utility:undeprecate,utility:underline,utility:undo,utility:unlinked,utility:unlock,utility:unmuted,utility:up,utility:upload,utility:user_role,utility:user,utility:variable,utility:video,utility:voicemail_drop,utility:volume_high,utility:volume_low,utility:volume_off,utility:waits,utility:warning,utility:watchlist,utility:weeklyview,utility:wifi,utility:work_order_type,utility:world,utility:yubi_key,utility:zoomin,utility:zoomout|}\" alternative-text=\"${2:Title}\" title=\"${2:Title}\" variant=\"${3|Success,warning,error,inverse|}\" size=\"${4|xx-small,x-small,small,medium,large|}\"></lightning-icon>",
                 ),
                 detail: "LWC Icon",
                 kind: vscode.CompletionItemKind.Snippet,
@@ -4093,14 +4274,6 @@ module.exports = {
     enableJavascriptSnippetsFlag,
     sendTeleEventforConfigs
 }
-
-
-
-// "Aura Checkbox": {
-//     "prefix": "adbn:checkbox-aura",
-//     "body": "",
-//     "description": ""
-// },
 
 // "LWC Checkbox":{
 //     "prefix": "adbn:checkbox-lwc",